# Improve bashlog.sh performance