Performance is improved by an order of magnitude by ensuring that
logging does not fork unnecessarily:
* forking to `date` only happens if timestamps are actually used
* forking to `basename` is replaced with bash expansion
* forking to `awk` for uppercasing is repaced with bash expansion

Fixes #196