@reegnz 2021-12-14
### Benchmarks

### Benchmarks
#### Before
```sh
❯ time ./terraform -version
Terraform v1.1.0
on darwin_amd64
./terraform -version  0.44s user 0.48s system 62% cpu 1.489 total
```
#### After
```sh
❯ time ./terraform -version
Terraform v1.1.0
on darwin_amd64
./terraform -version  0.07s user 0.04s system 78% cpu 0.147 total
```
I made sure that the log function stays backward compatible, so there shouldn't be any breaking changes.