@cspotcode 2022-03-07
I just tested this.  Here are some hyperfine results:

```
ubuntu:~/.tfenv$ hyperfine './versions/1.1.7/terraform version'
Benchmark 1: ./versions/1.1.7/terraform version
  Time (mean ± σ):      37.7 ms ±   9.4 ms    [User: 34.4 ms, System: 14.9 ms]
  Range (min … max):    27.2 ms …  69.6 ms    42 runs
 
ubuntu:~/.tfenv$ hyperfine './bin/terraform version'
Benchmark 1: ./bin/terraform version
  Time (mean ± σ):      85.3 ms ±  14.6 ms    [User: 69.1 ms, System: 27.3 ms]
  Range (min … max):    66.6 ms … 120.9 ms    30 runs
```

Only ~45ms of overhead is pretty good.  I hope this gets merged!