@reegnz 2022-03-08
@cspotcode I think we could use BASHLOG_ENABLE=1 otherwise it should be a noop. The `terraform` wrapper should default to disabled (so noop), unless the automatic installation happens, then we'd most definitely want to see what's going on.