@cspotcode 2022-03-08
How about `TFENV_BASHLOG_ENABLE` to namespace it the same way as the other environment variables?

We can declare a shim `enable_full_bashlog` that can be called in the auto-install codepaths to switch to the full bashlog immediately before doing stuff that needs to be logged.