@cspotcode 2022-03-09
I realized I did not explain my implementation.

I noticed that `log debug ...` is a no-op unless you have debugging enabled via log level.  And the terraform shim only does debug logging in the common case where you already have terraform installed.

So I changed helpers.sh to declare a `function log` shim that will lazy-load the full bashlog as soon as non-debug logging happens.  If you have debug logging enabled, it does the right thing and loads bashlog immediately.

This avoids the need for any new environment variables, so it's not a behavioral change.  Things should log in the same way they always have.

I also moved some of the `libexec` commands into `lib` as functions, which avoids a few spawns.