@reegnz 2022-03-09
@cspotcode that is awesome!
I did some additional perf tweaking (getting rid of forking to basename and dirname in the hot path), it added another significant bump in perf:

Before tweaking:
```sh
❯ hyperfine -w 10 '~/.cache/terraform/versions/1.1.5/terraform version' './terraform version'
Benchmark 1: ~/.cache/terraform/versions/1.1.5/terraform version
  Time (mean ± σ):      36.3 ms ±   0.9 ms    [User: 31.7 ms, System: 13.5 ms]
  Range (min … max):    34.5 ms …  38.7 ms    67 runs

Benchmark 2: ./terraform version
  Time (mean ± σ):     105.8 ms ±   2.3 ms    [User: 50.8 ms, System: 36.8 ms]
  Range (min … max):   101.8 ms … 111.1 ms    25 runs

Summary
  '~/.cache/terraform/versions/1.1.5/terraform version' ran
    2.91 ± 0.10 times faster than './terraform version'
```
After the tweaks:

```sh
❯ hyperfine -w 10 '~/.cache/terraform/versions/1.1.5/terraform version' './terraform version'
Benchmark 1: ~/.cache/terraform/versions/1.1.5/terraform version
  Time (mean ± σ):      36.1 ms ±   1.0 ms    [User: 31.8 ms, System: 13.4 ms]
  Range (min … max):    34.0 ms …  39.8 ms    68 runs

Benchmark 2: ./terraform version
  Time (mean ± σ):      73.8 ms ±   1.5 ms    [User: 44.1 ms, System: 25.8 ms]
  Range (min … max):    70.9 ms …  77.6 ms    36 runs

Summary
  '~/.cache/terraform/versions/1.1.5/terraform version' ran
    2.05 ± 0.07 times faster than './terraform version'
```

This PR starts to look really good now!