@pinkplus 2022-03-14
The idea of lazy loading `log()` is really clever. I've confirmed that it works with Bash 3 as well and is super fast. Thanks!