@cspotcode 2022-03-14
Excellent, thanks for confirming!

Is there anything else Zordrak needs to merge this?  Looks like automated tests run on both Mac and Linux, so I think we've covered all our bases.