@RoryKiefer 2022-03-16
Many thanks for tackling these issues, cspot/reegnz! ❤️ This is a fantastic performance improvement.