@cspotcode 2022-05-04
Bump!  Can we at least get this issue or pull request pinned?  That way, anyone who hits similar performance issues will be able to find this branch and use it.