@cspotcode 2022-07-07
In the interim, the git installation method works well.